pyproject.toml
[project]
name = "pymccool"
dynamic = ["version"]
authors = [
  { name="Brendon McCool", email="brendonmccool@gmail.com" },
]
description = "Reusable Python Utilities by McCool"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "colorlog",
    "mock",
    "opentelemetry-exporter-otlp-proto-grpc",
    "opentelemetry-exporter-otlp-proto-http",
    "python-logging-loki"
]

[build-system]
requires = ["setuptools", "setuptools_scm[toml]>=6.2", "python-semantic-release"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

[tool.semantic_release]
version_variable = "pyproject.toml:version" # version location
branch = "main"                             # branch to make releases of
changelog_file = "CHANGELOG.md"             # changelog file
upload_to_release = true                    # auto-create GitHub release
upload_to_pypi = false                      # don't auto-upload to PyPI
patch_without_tag = true                    # patch release by default

[project.urls]
"Homepage" = "https://github.com/bmccool/pyMcCool"

[tool.yapf]
based_on_style = "pep8"
spaces_before_comment = 4

[tool.pytest.ini_options]
markers = [
    "e2e: End-to-End, requires tracing and logging setup, not meant to run as unit tests (deselect with '-m \"not e2e\"')"
]
